refactor(minimax-h3): remove dead and duplicate kernels
